refactor(dirs): moved test files to __tests__ folders
