+++ app/lib/plugins/settings/__tests__/get.spec.js
@@ -1,4 +1,4 @@
-import getUserSettings from './get'
+import getUserSettings from '../get'
 
 const plugin = {
   settings: {
